chore: move `tests::shared` items related to filesystem operations into the submodule
