+++ src/models/blockchain/transaction/validity/single_proof.rs
@@ -1003,10 +1003,10 @@ mod tests {
     }
 
     mod update_tests {
+        use proptest::prelude::*;
         use rand::random;
         use tasm_lib::hashing::merkle_verify::MerkleVerify;
         use twenty_first::prelude::Mmr;
-        use proptest::prelude::*;
 
         use crate::models::blockchain::transaction::transaction_kernel::TransactionKernelModifier;
         use crate::models::blockchain::transaction::validity::tasm::single_proof::update_branch::tests::deterministic_update_witness_additions_and_removals;
