+++ src/rpc_auth.rs
@@ -284,7 +284,7 @@ mod tests {
     use macro_rules_attr::apply;
 
     use super::*;
-    use crate::tests::shared::unit_test_data_directory;
+    use crate::tests::shared::files::unit_test_data_directory;
     use crate::tests::shared_tokio_runtime;
 
     mod token {
